refactor: convert `apps/meteor/app/apps/server/converters/visitors.js` to TypeScript
